cli: add fpgad_api crate which can be shared with users to give access to dbus calls
this hopefully gives us a way to share the interface changes in a maintainable way
