+++ cli/src/main.rs
@@ -10,8 +10,6 @@
 //
 // You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
 
-mod proxies;
-
 mod status;
 
 mod load;
