+++ fpgad_api/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod control;
+pub mod proxies;
+pub mod status;
