+++ fpgad_api/src/proxies/control_proxy.rs
