+++ fpgad_api/src/proxies/status_proxy.rs
